requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.7.0
pillow>=9.0.0
numpy==1.26.4
huggingface-hub>=0.14.0
tqdm>=4.65.0
pytest>=7.0.0
